refactor: simplify guardian switch authentication
