+++ src/components/GuardianTransitionHero.tsx
