+++ src/lib/intercom/desktop-adapter.test.ts
@@ -13,7 +13,6 @@ jest.mock('lib/miden/back/actions', () => ({
   registerNewWallet: jest.fn().mockResolvedValue(undefined),
   registerImportedWallet: jest.fn().mockResolvedValue(undefined),
   unlock: jest.fn().mockResolvedValue(undefined),
-  reauthenticate: jest.fn().mockResolvedValue(undefined),
   lock: jest.fn().mockResolvedValue(undefined),
   createHDAccount: jest.fn().mockResolvedValue(undefined),
   updateCurrentAccount: jest.fn().mockResolvedValue(undefined),
@@ -127,13 +126,6 @@ describe('DesktopIntercomAdapter', () => {
 
       expect(Actions.unlock).toHaveBeenCalledWith('test123');
       expect(response).toEqual({ type: WalletMessageType.UnlockResponse });
-    });
-
-    it('handles hardware ReauthenticateRequest without returning secrets', async () => {
-      const response = await adapter.request({ type: WalletMessageType.ReauthenticateRequest });
-
-      expect(Actions.reauthenticate).toHaveBeenCalledWith(undefined);
-      expect(response).toEqual({ type: WalletMessageType.ReauthenticateResponse });
     });
 
     it('handles LockRequest', async () => {
