+++ src/lib/store/types.ts
@@ -136,7 +136,6 @@ export interface WalletActions {
     walletAccounts: WalletAccount[]
   ) => Promise<void>;
   unlock: (password?: string) => Promise<void>;
-  reauthenticate: (password?: string) => Promise<void>;
 
   // Account actions
   createAccount: (walletType: WalletType, name?: string) => Promise<void>;
